Fixed typescript errors
