+++ frontend/src/services/shields.ts
@@ -1,9 +1,8 @@
 import { SHIELDS_API_ENDPOINT } from '@/config/api';
+import { Shield } from '@/types/auth';
 
-export interface Shield {
-  identifier: string;
-  name?: string;
-}
+// Re-export types for backward compatibility
+export type { Shield } from '@/types/auth';
 
 interface BackendShield {
   id: string;
