+++ frontend/src/types/api.ts
@@ -1,0 +1,32 @@
+export interface Model {
+  model_name: string;
+  provider_resource_id: string;
+  model_type: string;
+}
+
+export interface EmbeddingModel {
+  name: string;
+  provider_resource_id: string;
+  model_type: string;
+}
+
+export interface Provider {
+  provider_id: string;
+  provider_type: string;
+  config: Record<string, unknown>;
+  api: string;
+}
+
+export interface LlamaStackResponse {
+  type: string;
+  content: string;
+  sessionId?: string;
+  tool?: {
+    name: string;
+    params?: Record<string, unknown>;
+  };
+}
+
+export type LlamaStackParser = {
+  parse(text: string): string | null;
+};
