+++ frontend/src/types/index.ts
@@ -1,0 +1,8 @@
+// Re-export all types for easier imports
+export * from './agent';
+export * from './auth';
+export * from './chat';
+export * from './forms';
+export * from './knowledge-base';
+export * from './tools';
+export * from './api';
